-add textSymbol to markers

+++ android/src/main/kotlin/com/valentingrigorean/arcgis_maps_flutter/convert/mapping/symbology/MarkerSymbolConvert.kt
@@ -4,7 +4,7 @@ import com.arcgismaps.mapping.symbology.MarkerSymbol
 
 fun MarkerSymbol.interpretMarkerSymbol(data: Map<*, *>) {
     data["angle"]?.let {
-        angle = it as Float
+        angle = (it as Double).toFloat()
     }
     data["angleAlignment"]?.let {
         angleAlignment = (it as String).toSymbolAngleAlignment()
@@ -12,13 +12,13 @@ fun MarkerSymbol.interpretMarkerSymbol(data: Map<*, *>) {
 
     data["leaderOffset"]?.let {
         val arr = it as List<*>
-        leaderOffsetX = arr[0] as Float
-        leaderOffsetY = arr[1] as Float
+        leaderOffsetX = (arr[0] as Double).toFloat()
+        leaderOffsetY = (arr[1] as Double).toFloat()
     }
 
     data["offset"]?.let {
         val arr = it as List<*>
-        offsetX = arr[0] as Float
-        offsetY = arr[1] as Float
+        offsetX = (arr[0] as Double).toFloat()
+        offsetY = (arr[1] as Double).toFloat()
     }
 }
